Version 0.0.4

+++ extensions/cql/src/cql/datetime.cpp
@@ -157,9 +157,9 @@ std::string DateTimeValue::to_string() const {
 
 int64_t DateTimeValue::to_julian_day() const {
 	// Julian Day Number calculation
-	int a = (14 - month) / 12;
-	int y = year + 4800 - a;
-	int m = month + 12 * a - 3;
+	int64_t a = (14 - month) / 12;
+	int64_t y = year + 4800 - a;
+	int64_t m = month + 12 * a - 3;
 	return day + (153 * m + 2) / 5 + 365 * y + y / 4 - y / 100 + y / 400 - 32045;
 }
